refactor: move compiling to codegen/mod.rs
