--- src/codegen.rs
@@ -1,4 +1,0 @@
-pub mod context;
-pub(crate) mod operations;
-mod pass_manager;
-pub use pass_manager::run_pass_manager;
